Attempt extract to package
