+++ package.json
@@ -6,6 +6,7 @@
     "packages/ui",
     "packages/docs",
     "packages/nodejs",
+    "packages/hjson",
     "packages/openai",
     "packages/private/*"
   ],
