+++ packages/hjson/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@data-story/hjson",
+  "version": "0.0.2",
+  "packageManager": "yarn@4.0.1",
+  "main": "dist/hjson.bundle.js",
+  "scripts": {
+    "build": "webpack --config webpack.config.js"
+  },
+  "type": "module",
+  "devDependencies": {
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
