+++ packages/hjson/src/hjson-common.js
