+++ packages/hjson/src/hjson-version.js
