+++ packages/hjson/src/index.d.ts
@@ -1,0 +1,5 @@
+// hjson.d.ts
+declare module '@data-story/hjson' {
+  const hjson: any;
+  export default hjson;
+}
